Item Info Widget & C++ Remove Item Function
With this push, you have the option to use an Inventory Item Info Widget for your inventory system. This is displayed on the right side of the inventory. If no valid item is selected, the widget switcher changes and a text appears that says “No Item selected.” In addition, the Remove Item Function has been rewritten in C++. The Static Mesh and Skeletal Mesh references in the Inventory Item have also been rewritten in C++, giving us more control over them.